+++ packages/core/src/index.mjs
@@ -44,7 +44,7 @@ const RENDERED_ELEMENTS = [
   "datalist", "dd", "del", "details", "dfn", "dialog", "div", "dl", "dt", "em", "embed",
   "fieldset", "figcaption", "figure", "footer", "form", "h1", "h2", "h3", "h4", "h5",
   "h6", "header", "hgroup", "hr", "i", "iframe", "img", "input", "ins", "kbd", "label",
-  "legend", "li", "main", "map", "mark", "menu", "meter", "nav", "object", "ol", "optgroup",
+  "legend", "li", "main", "map", "mark", "math", "menu", "meter", "nav", "object", "ol", "optgroup",
   "option", "output", "p", "picture", "pre", "progress", "q", "rp", "rt", "ruby", "s",
   "samp", "section", "select", "slot", "small", "source", "span", "strong", "sub", "summary",
   "sup", "table", "tbody", "td", "textarea", "tfoot", "th", "thead", "time", "tr", "track",